refactor: convert Content to functional component and move schema validator to its own file #112
